+++ cms-20240330/alibabacloud_cms20240330/models/_create_prometheus_virtual_instance_request.py
@@ -8,11 +8,13 @@ class CreatePrometheusVirtualInstanceRequest(DaraModel):
     def __init__(
         self,
         namespace: str = None,
+        tenant_id: str = None,
     ):
         # Each cloud product can only create one virtual instance in each region.
         # 
         # This parameter is required.
         self.namespace = namespace
+        self.tenant_id = tenant_id
 
     def validate(self):
         pass
@@ -25,12 +27,18 @@ class CreatePrometheusVirtualInstanceRequest(DaraModel):
         if self.namespace is not None:
             result['namespace'] = self.namespace
 
+        if self.tenant_id is not None:
+            result['tenantId'] = self.tenant_id
+
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('namespace') is not None:
             self.namespace = m.get('namespace')
+
+        if m.get('tenantId') is not None:
+            self.tenant_id = m.get('tenantId')
 
         return self